Refactoring metadata components for future extension
